+++ frontend/src/actionCreators/cartActionCreators.js
@@ -5,7 +5,6 @@ import { CART_ADD_ITEM, CART_REMOVE_ITEM } from "../constants/cartConstants";
 import { URL } from "../constants/urlConstants";
 
 export const addToCart = (id, qty) => async (dispatch, getState) => {
-  // const { data } = await axios.get(`https://fakestoreapi.com/products/${id}`);
   const { data } = await axios.get(`${URL}/api/products/${id}`);
   dispatch({
     type: CART_ADD_ITEM,
@@ -21,7 +20,6 @@ export const addToCart = (id, qty) => async (dispatch, getState) => {
   });
 
   localStorage.setItem("cartItems", JSON.stringify(getState().cart.cartItems));
-  // console.log(getState().cart.cartItems)
 };
 
 export const removeFromCart = (id) => (dispatch, getState) => {
